backend/requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
python-dotenv==1.0.1